Support 12.0

+++ SpellWork/DBC/Structures/AreaTableEntry.cs
@@ -21,8 +21,8 @@ namespace SpellWork.DBC.Structures
         public uint UwIntroSound;
         public byte FactionGroupMask;
         public float AmbientMultiplier;
-        public byte MountFlags;
-        public short PvpCombatWorldStateID;
+        public int MountFlags;
+        public int PvpCombatWorldStateID;
         public byte WildBattlePetLevelMin;
         public byte WildBattlePetLevelMax;
         public byte WindSettingsID;
